In case a thread is doing a blocking operation (e.g. read()) on a serial device, while it is being terminated by pthread_cancel(), then uart_close() gets called, but the semaphore (dev->recv.sem in the above example) is still blocked.
This means that once the serial device is opened next time, data will
arrive on the serial port (and driver interrupts handled as normal), but
the received characters never arrive in the reader thread.

This patch addresses the problem by re-initializing the semaphores on the
last uart_close() on the device.