stm32f429i-disco: add default platform configuration to test framebuffer support for the ltdc
Memory configuration:

The heap start address of memory region 3 starts at 0xD0000000.
The allocated framebuffer of the layers is addressed up to the end of the memory
region 3.

If the layer pixel format changes e.g. from RGB565 to RGB24, than the memory
configuration must be configured too. Currently an easy memory calculation
is missing for that in Kconfig.

Signed-off-by: Marco Krahl <ocram.lhark@gmail.com>