This change fixes a build problem that only occurs when reconfiguring from Linux to Windows or vice-versa.  It is a problem that was present but not usually experienced until two things happened:  (1) The pre_config target was added to run before the menconfig operation and (2) the context target was added before the pre_config target in order to set up the correct symbolic links (in the apps/platform directory) needed by the pre_config target.
But then now if you start with a Linux system and run 'make menuconfig' to switch to Linux, the context target will execute first and set up POSIX style symbolic links before doing the menuconfig.  Then after the menuconfig, the make will fail on Windows if you are using a Windows native toolchain because that native toolchain cannot follow the Cygwin-style symbolic links.

The first here is to also execute the clean_context AFTER executing menuconfig.  A lot more happens now:  It used to be that doing 'make menuconfig' only did the menuconfig operation.  No it does context, pre_config, menuconfig, clean_context.  Not nearly as snappy as it used to be.