STMF7xxx RTC  Remove proxy #defines  Ensure the LSE(ON) etal are set and remembered in   a) A cold start (RTC_MAGIC invalid) of the RTC   b) A warm start (RTC_MAGIC valid) of the RTC but a clock change.
The change was needed because in bench testing a merge of the latest's STM32  53ec3ca (and friends) it became apparent that the
Sequence of operation is wrong in the reset of the Backup Domain in the RCC code.  PWR is required before the Backup Domain
can be futzed with. !!!This Code should be tested on STM32 and if needed rippled to the STM32 families