update / correct some compilation and linkage settings
HOSTCFLAGS: Should not be set the same as CFLAGS since CFLAGS may be for a completely different platform.  Use sane defaults like other configs/*/*/Make.defs.

ARCHCPUFLAGS: Depending on how the GNU compiler was built we may need -fno-stack-protector.  Since we always want it we can just default to specifying it instead of depending on the compiler default.

LDFLAGS: If compiling for a 32 bit target from a 64 bit platform we need to tell the linker to generate a 32 bit elf file with -m elf_i386.