Top-Level Makefiles.  Fix a chicken-and-egg problem.
In the menuconfig target, the context dependency was executed before kconfig-mconf.  That was necessary because the link at apps/platform/board needed to be set up before creating the apps/Kconfig file.  Otherwise, the platform Kconfig files would not be included.  But this introduces the chicken-and-egg problem in some configurations.

In particular:  (1) An NX graphics configuration is used that requires auto-generation of source files using cpp, (2) the configuration is set for Linux, but (3) we are running under Cygwin with (4) a Windows native toolchain.  In this case, POSIX-style symbolic links are set up but the Windows native toolchain cannot follow them.

The reason we are running 'make menuconfig' is to change from Linux to Cygwin, but the target fails.  During the context phase, NX runs CPP to generate source files but that fails because the Windows native toolchain cannot follow the links.  Checkmate.

This was fixed by changing all of the make menuconfig (and related) targets.  They no long depend on context being run.  Instead, they depend only on the dirlinks target.  The dirlinks target only sets up the directory links but does not try to run all of the context setup; the compiler is never invoked; no code is autogeneraed; and things work.