Added initialization code and Kconfig entries to set up the LIS3DSH accelerometer driver on STM32F4Discovery rev. C boards and attach the associated interrupt callback.
Added the argument parameter (FAR void *arg) to the interrupt handler provided by the LIS3DSH
driver to fit the definition for ISRs in xcpt_t.

Changed the check for working queue availability in lis3dsh interrupt handler to use work_available()
and not crash in case of an overrun.